Significantly cleaned up inlining code and moved it to a new file.
From AstTransforms.cpp to InlineRelationsTransformer.cpp
