+++ src/AstTransforms.cpp
@@ -18,7 +18,6 @@
 #include "AstTypeAnalysis.h"
 #include "AstUtils.h"
 #include "AstVisitor.h"
-#include "MagicSet.h"
 #include "PrecedenceGraph.h"
 
 namespace souffle {
@@ -694,44 +693,9 @@ bool MaterializeAggregationQueriesTransformer::needsMaterializedRelation(const A
     return false;
 }
 
+// TODO (azreika): fix this up to only add when necessary
 bool EvaluateConstantAggregatesTransformer::transform(AstTranslationUnit& translationUnit) {
   // TODO: need some way of checking if things have changed
-
-  // TODO: commenting
-  // struct M : public AstNodeMapper {
-  //   M() {}
-  //   std::unique_ptr<AstNode> operator()(std::unique_ptr<AstNode> node) const override {
-  //     if(AstAggregator* aggr = dynamic_cast<AstAggregator*>(node.get())) {
-  //       AstAggregator::Op op = aggr->getOperator();
-  //       if(op == AstAggregator::min || op == AstAggregator::max) {
-  //         const AstArgument* arg = aggr->getTargetExpression();
-  //         if (arg != nullptr && dynamic_cast<const AstConstant*>(arg)) {
-  //           return std::unique_ptr<AstNode>(arg->clone());
-  //         }
-  //       } else if (op == AstAggregator::count) {
-  //         // TODO: check correctness of this part!!! Mainly:
-  //         //        - does this check correctly that there is no body atom?
-  //         //        - should it always be 1? could be 0? in inlining's case, yes
-  //         if (!dynamic_cast<const AstAtom*>(aggr->getBodyLiterals()[0])) {
-  //           return std::unique_ptr<AstNode>(new AstNumberConstant(1));
-  //         }
-  //       } else if (op == AstAggregator::sum) {
-  //         // TODO: Same as above
-  //         const AstArgument* arg = aggr->getTargetExpression();
-  //         if (arg != nullptr && dynamic_cast<const AstConstant*>(arg)) {
-  //           return std::unique_ptr<AstNode>(arg->clone());
-  //         }
-  //       } else {
-  //         // TODO: correctness?
-  //         assert(false && "unhandled aggregate operator type");
-  //       }
-  //     }
-  //
-  //     node->apply(*this);
-  //     return node;
-  //   }
-  // };
-
   AstRelation* trueRel = new AstRelation();
   trueRel->setName("+true+");
   AstClause* fact = new AstClause();
@@ -861,933 +825,6 @@ bool RemoveRedundantRelationsTransformer::transform(AstTranslationUnit& translat
         }
     }
     return changed;
-}
-
-// TODO: not really Martelli-Montanari algo anymore?
-// Reduces a vector of substitutions based on Martelli-Montanari algorithm
-bool reduceSubstitution(std::vector<std::pair<AstArgument*, AstArgument*>>& sub) {
-  // Type-Checking functions
-  // TODO: is [&] needed?
-  auto isConstant = [&](AstArgument* arg) {return (dynamic_cast<AstConstant*>(arg));};
-  auto isRecord = [&](AstArgument* arg) {return (dynamic_cast<AstRecordInit*>(arg));};
-  auto isCounter = [&](AstArgument* arg) {return (dynamic_cast<AstCounter*>(arg));};
-
-  // Keep trying to reduce the substitutions until we reach a fixed point
-  bool done = false;
-  while (!done) {
-    done = true;
-
-    // Try reducing each pair
-    for(int i = 0; i < sub.size(); i++) {
-      auto currPair = sub[i];
-      AstArgument* lhs = currPair.first;
-      AstArgument* rhs = currPair.second;
-
-      // TODO: see doubles_new.dl file for what's left
-      // AstArgument Cases:
-      //  0 - AstVariable
-      //  1 - AstUnnamedVariable (_) -> change all to variables to reduce this case -> need to rename everything
-      //  3 - AstConstant -> as straightforward as variables
-      //  4 - AstFunctor -> can this just be direct? or more thought into this?
-      //  5 - AstRecordInit -> this should just be pairwise equality right? think about this...
-      //  6 - AstTypeCast -> find out what this does ;; SIDE NOTE: no idea how this looks actually but should work immediately
-      //  7 - AstAggregator -> complicated! think about this - perhaps equate arguments on a lower level
-
-      // TODO: throw this error properly
-      assert(!isCounter(lhs) && !isCounter(rhs) && "ERROR: CONTAINS `$`");
-
-      if (*lhs == *rhs) {
-        // get rid of `x = x`
-        sub.erase(sub.begin() + i);
-        done = false;
-      } else if (isConstant(lhs) && isConstant(rhs)) {
-        // both are constants but not equal (prev case => !=)
-        // failed to unify!
-        return false;
-      } else if (isRecord(lhs) && isRecord(rhs)) {
-        // handle equated records
-        // note: we will not deal with the case where only one side is
-        // a record and the other is a variable
-        // TODO: think about the case where record = constant, etc.
-        std::vector<AstArgument*> lhsArgs = static_cast<AstRecordInit*>(lhs)->getArguments();
-        std::vector<AstArgument*> rhsArgs = static_cast<AstRecordInit*>(rhs)->getArguments();
-        if (lhsArgs.size() != rhsArgs.size()) {
-          return false;
-        }
-        for (int i = 0; i < lhsArgs.size(); i++) {
-          sub.push_back(std::make_pair(lhsArgs[i], rhsArgs[i]));
-        }
-        sub.erase(sub.begin() + i);
-        done = false;
-      } else if ((isRecord(lhs) && isConstant(rhs))||(isConstant(lhs) && isRecord(rhs))) {
-        // a record =/= a constant
-        // TODO: is this correct? is this even possible?
-        return false;
-      }
-    }
-  }
-
-  return true;
-}
-
-// Returns the pair (v, p), where p indicates whether unification has failed,
-// and v is the vector of substitutions to unify the two given atoms if successful.
-// Assumes that the atoms are both of the same relation.
-std::pair<std::vector<std::pair<AstArgument*, AstArgument*>>, bool> unifyAtoms(AstAtom* first, AstAtom* second) {
-  std::pair<std::vector<std::pair<AstArgument*, AstArgument*>>, bool> result;
-  std::vector<std::pair<AstArgument*, AstArgument*>> substitution;
-
-  std::vector<AstArgument*> firstArgs = first->getArguments();
-  std::vector<AstArgument*> secondArgs = second->getArguments();
-
-  // Create the initial unification equalities
-  for(int i = 0; i < firstArgs.size(); i++) {
-    substitution.push_back(std::make_pair(firstArgs[i], secondArgs[i]));
-  }
-
-  // Reduce the substitutions
-  bool success = reduceSubstitution(substitution);
-
-  result = make_pair(substitution, success);
-  return result;
-}
-
-std::pair<std::vector<AstLiteral*>, bool> inlineBodyLiterals(AstAtom* atom, AstClause* atomClause) {
-  // NOTE: this renaming is done to fix the problem when an inined atom is used twice in a body
-  static int inlineCount = 0;
-  bool changed = false;
-  std::vector<AstLiteral*> addedLits;
-
-  struct VariableRenamer : public AstNodeMapper {
-    int varnum;
-    VariableRenamer(int varnum) : varnum(varnum) {}
-    std::unique_ptr<AstNode> operator()(std::unique_ptr<AstNode> node) const override {
-        if(AstVariable* var = dynamic_cast<AstVariable*>(node.get())){
-          // Rename the variable
-          AstVariable* newVar = var->clone();
-          std::stringstream newName;
-          // TODO: the var name can get way too long
-          newName << var->getName() << "_" << varnum;
-          newVar->setName(newName.str());
-          inlineCount++;
-          return std::unique_ptr<AstNode>(newVar);
-        }
-        node->apply(*this);
-        return node;
-    }
-  };
-
-  VariableRenamer update(inlineCount);
-  inlineCount++;
-  atomClause->apply(update);
-
-  // Get the constraints needed to unify the two atoms
-  std::pair<std::vector<std::pair<AstArgument*, AstArgument*>>, bool> res = unifyAtoms(atomClause->getHead(), atom);
-  if(res.second) {
-    changed = true;
-    for(auto pair : res.first) {
-      addedLits.push_back(new AstConstraint(BinaryConstraintOp::EQ,
-              std::unique_ptr<AstArgument>(pair.first->clone()),
-              std::unique_ptr<AstArgument>(pair.second->clone())));
-    }
-
-    // Add in the body of the current clause of the inlined atom
-    for(AstLiteral* lit : atomClause->getBodyLiterals()) {
-      addedLits.push_back(lit->clone());
-    }
-  }
-
-  return std::make_pair(addedLits, changed);
-}
-
-std::pair<std::vector<AstLiteral*>, bool> inlineNegatedBodyLiterals (AstAtom* atom, AstClause* atomClause, std::vector<AstConstraint*>& constraints) {
-  // NOTE: this renaming is done to fix the problem when an inined atom is used twice in a body
-  static int inlineCount = 0;
-  bool changed = false;
-  std::vector<AstLiteral*> addedLits;
-
-  struct VariableRenamer : public AstNodeMapper {
-    int varnum;
-    VariableRenamer(int varnum) : varnum(varnum) {}
-    std::unique_ptr<AstNode> operator()(std::unique_ptr<AstNode> node) const override {
-        if(AstVariable* var = dynamic_cast<AstVariable*>(node.get())){
-          // Rename the variable
-          AstVariable* newVar = var->clone();
-          std::stringstream newName;
-          // TODO: the var name can get way too long
-          newName << var->getName() << "_" << varnum;
-          newVar->setName(newName.str());
-          inlineCount++;
-          return std::unique_ptr<AstNode>(newVar);
-        }
-        node->apply(*this);
-        return node;
-    }
-  };
-
-  VariableRenamer update(inlineCount);
-  inlineCount++;
-  atomClause->apply(update);
-
-  // Get the constraints needed to unify the two atoms
-  std::pair<std::vector<std::pair<AstArgument*, AstArgument*>>, bool> res = unifyAtoms(atomClause->getHead(), atom);
-  if(res.second) {
-    changed = true;
-    for(auto pair : res.first) {
-      constraints.push_back(new AstConstraint(BinaryConstraintOp::EQ,
-              std::unique_ptr<AstArgument>(pair.first->clone()),
-              std::unique_ptr<AstArgument>(pair.second->clone())));
-    }
-
-    // Add in the body of the current clause of the inlined atom
-    for(AstLiteral* lit : atomClause->getBodyLiterals()) {
-      addedLits.push_back(lit->clone());
-    }
-  }
-
-  return std::make_pair(addedLits, changed);
-}
-
-// Renames all arguments contained in inlined relations to prevent conflict
-void renameInlinedArguments(AstProgram& program) {
-  // construct mapping to rename arguments
-  struct VariableRenamer : public AstNodeMapper {
-      VariableRenamer() {}
-      std::unique_ptr<AstNode> operator()(std::unique_ptr<AstNode> node) const override {
-          if(AstVariable* var = dynamic_cast<AstVariable*>(node.get())){
-            // Rename the variable
-            AstVariable* newVar = var->clone();
-            newVar->setName("<inlined_" + var->getName() + ">");
-            return std::unique_ptr<AstNode>(newVar);
-          }
-          node->apply(*this);
-          return node;
-      }
-  };
-
-  // TODO: fix this so that it fits the proper style?
-  VariableRenamer renameVar;
-  for(AstRelation* rel : program.getRelations()) {
-    if(rel->isInline()) {
-      for(AstClause* clause : rel->getClauses()) {
-        clause->apply(renameVar);
-      }
-    }
-  }
-}
-
-// Removes all underscores in inlined relations
-void nameUnderscores(AstProgram& program) {
-  struct M : public AstNodeMapper {
-    M() {}
-    std::unique_ptr<AstNode> operator()(std::unique_ptr<AstNode> node) const override {
-      static int underscoreCount = 0;
-      if (AstUnnamedVariable* var = dynamic_cast<AstUnnamedVariable*>(node.get())) {
-        // Give a unique name to the underscored vairable
-        std::stringstream newVarName;
-        // TODO: match NormaliseConstraintsTransformer for consistency?
-        newVarName << "<underscore_" << underscoreCount++ << ">";
-        AstVariable* newVar = new AstVariable(newVarName.str());
-        return std::unique_ptr<AstNode>(newVar);
-      }
-      node->apply(*this);
-      return node;
-    }
-  };
-
-  M update;
-  program.apply(update);
-}
-
-// Replaces all previously unnamed variables back with underscores
-void unnameUnderscores(AstProgram& program) {
-  struct M : public AstNodeMapper {
-    M() {}
-    std::unique_ptr<AstNode> operator()(std::unique_ptr<AstNode> node) const override {
-      if(AstVariable* var = dynamic_cast<AstVariable*>(node.get())) {
-        std::string name = var->getName();
-        if((name.find("<underscore_")==0) && (name[name.length()-1] == '>')) {
-          return std::unique_ptr<AstNode>(new AstUnnamedVariable());
-        }
-      }
-      node->apply(*this);
-      return node;
-    }
-  };
-
-  M update;
-  program.apply(update);
-}
-
-// Replaces all previously unnamed variables back with underscores
-void renameUnderscores(AstArgument* arg) {
-  struct M : public AstNodeMapper {
-    M() {}
-    std::unique_ptr<AstNode> operator()(std::unique_ptr<AstNode> node) const override {
-      static int underscoreCount = 0;
-      if(AstVariable* var = dynamic_cast<AstVariable*>(node.get())) {
-        std::string name = var->getName();
-        if((name.find("<underscore_")==0) && (name[name.length()-1] == '>')) {
-          // Give a unique name to the underscored vairable
-          std::stringstream newVarName;
-          // TODO: match NormaliseConstraintsTransformer for consistency?
-          newVarName << "<underscore_" << underscoreCount++ << ">";
-          AstVariable* newVar = new AstVariable(newVarName.str());
-          return std::unique_ptr<AstNode>(newVar);        }
-      }
-      node->apply(*this);
-      return node;
-    }
-  };
-
-  M update;
-  arg->apply(update);
-}
-
-// TODO: move this into renameTargetExpressions
-std::string getNewAggrTargetName(std::string name, int varnum) {
-    // TODO: should match <A_N>,  where N is a number
-    int nameLength = name.length();
-    int lastUnderscore = name.find_last_of('_');
-
-    std::stringstream newVarName;
-    if(nameLength > 0 && name[0] == '<' && name[nameLength-1] == '>' && lastUnderscore != name.npos){
-        newVarName << name.substr(0, lastUnderscore) << "_" << varnum << ">";
-    } else {
-        newVarName << "<" << name << "_" << varnum << ">";
-    }
-    return newVarName.str();
-}
-
-// TODO: this is only necessary because for some reason the aggregators are sharing variable scope
-// Replaces all previously unnamed variables back with underscores
-void renameTargetExpressions(AstArgument* arg) {
-  renameUnderscores(arg);
-  struct N : public AstNodeMapper {
-    std::string name;
-    int varnum;
-    N(std::string name, int varnum) : name(name), varnum(varnum) {}
-    std::unique_ptr<AstNode> operator()(std::unique_ptr<AstNode> node) const override {
-      if(AstVariable* var = dynamic_cast<AstVariable*>(node.get())) {
-        if(var->getName() == name) {
-          // Give a unique new name to the variable
-          // TODO: match NormaliseConstraintsTransformer for consistency?
-          std::string newVarName = getNewAggrTargetName(name, varnum);
-          AstVariable* newVar = new AstVariable(newVarName);
-          return std::unique_ptr<AstNode>(newVar);
-        }
-      }
-      node->apply(*this);
-      return node;
-    }
-  };
-
-  struct M : public AstNodeMapper {
-    M() {}
-    std::unique_ptr<AstNode> operator()(std::unique_ptr<AstNode> node) const override {
-      static int targetExpr = 0;
-      if(AstAggregator* aggr = dynamic_cast<AstAggregator*>(node.get())) {
-        AstAggregator* newAggr = aggr->clone();
-        const AstArgument* target = newAggr->getTargetExpression();
-        if(target == nullptr) {
-          node->apply(*this);
-          return node;
-        }
-        if(const AstVariable* var = dynamic_cast<const AstVariable*>(target)) {
-          N update(var->getName(), targetExpr++);
-          newAggr->apply(update);
-        }
-        newAggr->apply(*this);
-        return std::unique_ptr<AstNode>(newAggr);
-      }
-      node->apply(*this);
-      return node;
-    }
-  };
-
-  M update;
-  arg->apply(update);
-}
-
-AstLiteral* negateLiteral(AstLiteral* lit) {
-  if(AstAtom* atom = dynamic_cast<AstAtom*>(lit)) {
-    AstNegation* neg = new AstNegation(std::unique_ptr<AstAtom>(atom->clone()));
-    return neg;
-  } else if (AstNegation* neg = dynamic_cast<AstNegation*>(lit)) {
-    AstAtom* atom = neg->getAtom()->clone();
-    return atom;
-  } else if (AstConstraint* cons = dynamic_cast<AstConstraint*>(lit)) {
-    AstConstraint* newCons = cons->clone();
-    newCons->negate();
-    return newCons;
-  } else {
-    assert(false && "this should've never been reached");
-    return lit; // TODO: THIS SHOULD NEVER BE REACHED
-  }
-}
-
-  std::vector<std::vector<AstLiteral*>> combineNegatedLiterals(std::vector<std::vector<AstLiteral*>> litGroups) {
-    std::vector<std::vector<AstLiteral*>> res;
-
-    std::vector<AstLiteral*> litGroup = litGroups[0];
-    if(litGroups.size() == 1) {
-      // TODO: better code style for vector instantiation? is this even correct vector usage?
-      for(int i = 0; i < litGroup.size(); i++) {
-        std::vector<AstLiteral*> currVec;
-        currVec.push_back(negateLiteral(litGroup[i]));
-        res.push_back(currVec);
-        // delete litGroup[i]; TODO: should i?
-      }
-      return res;
-    }
-
-    std::vector<std::vector<AstLiteral*>> combinedRHS = combineNegatedLiterals(std::vector<std::vector<AstLiteral*>>(litGroups.begin()+1, litGroups.end()));
-    for(AstLiteral* lhsLit : litGroup) {
-      for(std::vector<AstLiteral*> rhsVec : combinedRHS) {
-        std::vector<AstLiteral*> currVec;
-        currVec.push_back(negateLiteral(lhsLit));
-
-        for(AstLiteral* lit : rhsVec) {
-          currVec.push_back(lit);
-        }
-
-        res.push_back(currVec);
-      }
-    }
-
-    return res;
-  }
-
-std::vector<std::vector<AstLiteral*>> formNegatedLiterals(AstProgram& program, AstAtom* atom) {
-  std::vector<std::vector<AstLiteral*>> addedBodyLiterals;
-  std::vector<std::vector<AstConstraint*>> addedConstraints;
-  for (AstClause* inClause : program.getRelation(atom->getName())->getClauses()) {
-    std::vector<AstConstraint*> currCons;
-    // Form the replacement clause
-    std::pair<std::vector<AstLiteral*>, bool> replacementBodyLiterals = inlineNegatedBodyLiterals(atom, inClause, currCons);
-    if(!replacementBodyLiterals.second) {
-      // Failed to unify
-      continue;
-    }
-    addedBodyLiterals.push_back(replacementBodyLiterals.first);
-    addedConstraints.push_back(currCons);
-  }
-  addedBodyLiterals = combineNegatedLiterals(addedBodyLiterals);
-
-  std::vector<std::vector<AstLiteral*>> newBodyLiterals;
-  for(auto subvector : addedBodyLiterals) {
-    newBodyLiterals.push_back(subvector);
-  }
-
-  // Add in the constraints to all the body literals
-  for(int i = 0; i < addedBodyLiterals.size(); i++) {
-    std::vector<AstLiteral*> currentBody = addedBodyLiterals[i];
-
-    for(std::vector<AstConstraint*> constraintGroup : addedConstraints) {
-      for(AstConstraint* constraint : constraintGroup) {
-        newBodyLiterals[i].push_back(constraint->clone()); // need to clone?
-      }
-    }
-  }
-
-  return newBodyLiterals;
-}
-
-std::pair<std::vector<AstArgument*>, bool> getInlinedArgument(AstProgram& program, const AstArgument* arg);
-std::pair<std::vector<AstAtom*>, bool> getInlinedAtom(AstProgram& program, AstAtom& atom);
-std::pair<std::vector<AstLiteral*>, bool> getInlinedLiteral(AstProgram& program, AstLiteral& literal);
-
-// TODO: NON-ATOM LITERALS ARE NOT HANLDED CORRECTLY - CHECK THROUGH THIS (e.g. negations)
-// ARGUMENTS: <<<inline body literals versions to add>, <inlined?>> <<replacement literal versions> <replacements exist?>>>
-std::pair<std::pair<std::vector<std::vector<AstLiteral*>>, bool>, std::pair<std::vector<AstLiteral*>, bool>> getInlinedLiteral(AstProgram& program, AstLiteral* lit) {
-  bool inlined = false;
-  bool changed = false;
-  std::vector<std::vector<AstLiteral*>> addedBodyLiterals;
-  std::vector<AstLiteral*> versions;
-
-  if(AstAtom* atom = dynamic_cast<AstAtom*>(lit)) {
-    // atom
-
-    // Check if this atom is meant to be inlined
-    AstRelation* rel = program.getRelation(atom->getName());
-    if (rel->isInline()) {
-      // We found an atom in the clause that needs to be inlined!
-      // The clause needs to be replaced
-      inlined = true;
-
-      // N new clauses should be formed, where N is the number of clauses
-      // associated with the inlined relation
-      for (AstClause* inClause : rel->getClauses()) {
-        // Form the replacement clause
-        std::pair<std::vector<AstLiteral*>, bool> replacementBodyLiterals = inlineBodyLiterals(atom, inClause);
-        if(!replacementBodyLiterals.second) {
-          // Failed to unify
-          continue;
-        }
-        addedBodyLiterals.push_back(replacementBodyLiterals.first);
-      }
-    } else {
-      // check if its arguments need to be changed
-      std::pair<std::vector<AstAtom*>, bool> atomVersions = getInlinedAtom(program, *atom);
-      if(atomVersions.second) {
-        changed = true;
-        for(AstAtom* newAtom : atomVersions.first) {
-          versions.push_back(newAtom);
-        }
-      }
-    }
-  } else if (AstNegation* neg = dynamic_cast<AstNegation*>(lit)) {
-    // check if its arguments need to be changed
-    AstAtom* atom = neg->getAtom();
-    auto atomVersions = getInlinedLiteral(program, atom);
-    if(atomVersions.first.second) {
-      inlined = true;
-
-      // c(x) :- b(x), !a(x) ==> c(x) :- b(x), !a1(x), !a2(x).
-      // TODO: check the unified atoms whether you can `and` them together without problems
-      addedBodyLiterals = formNegatedLiterals(program, atom);
-    } else if (atomVersions.second.second) {
-      changed = true;
-      for(AstLiteral* nextLit : atomVersions.second.first) {
-        // TODO: is this astliteral return correct?
-        AstAtom* newAtom = dynamic_cast<AstAtom*>(nextLit);
-        assert("atom expected" && newAtom); // TODO: fix this up
-        AstLiteral* newNeg = new AstNegation(std::unique_ptr<AstAtom>(newAtom));
-        versions.push_back(newNeg);
-      }
-    }
-  } else {
-    // constraint
-    AstConstraint* constraint = dynamic_cast<AstConstraint*>(lit);
-    std::pair<std::vector<AstArgument*>, bool> lhsVersions = getInlinedArgument(program, constraint->getLHS());
-    if(lhsVersions.second) {
-      changed = true;
-      for(AstArgument* newLhs : lhsVersions.first) {
-        AstLiteral* newLit = new AstConstraint(constraint->getOperator(), std::unique_ptr<AstArgument>(newLhs), std::unique_ptr<AstArgument>(constraint->getRHS()->clone()));
-        versions.push_back(newLit);
-      }
-    } else {
-      std::pair<std::vector<AstArgument*>, bool> rhsVersions = getInlinedArgument(program, constraint->getRHS());
-      if(rhsVersions.second) {
-        changed = true;
-        for(AstArgument* newRhs : rhsVersions.first) {
-          AstLiteral* newLit = new AstConstraint(constraint->getOperator(), std::unique_ptr<AstArgument>(constraint->getLHS()->clone()), std::unique_ptr<AstArgument>(newRhs));
-          versions.push_back(newLit);
-        }
-      }
-    }
-  }
-
-  return std::make_pair(std::make_pair(addedBodyLiterals, inlined), std::make_pair(versions, changed));
-}
-
-AstArgument* combineAggregators(std::vector<AstArgument*> aggrs, BinaryOp fun) {
-  if(aggrs.size() == 1) {
-    return aggrs[0];
-  }
-
-  AstArgument* rhs = combineAggregators(std::vector<AstArgument*>(aggrs.begin() + 1, aggrs.end()), fun);
-
-  AstArgument* result = new AstBinaryFunctor(fun, std::unique_ptr<AstArgument>(aggrs[0]), std::unique_ptr<AstArgument> (rhs));
-
-  std::cout << "COMBINING AGGREGATORS" << std::endl;
-  for(auto agg : aggrs) {
-    if(agg == nullptr) {
-      std::cout << "WHAT?????" << std::endl;
-    } else {
-      std::cout << "LOOK: " << agg << std::endl;
-    }
-  }
-  std::cout << "GIVES US" << std::endl;
-  std::cout << *result << std::endl;
-  std::cout << "END" << std::endl;
-
-  renameTargetExpressions(result);
-
-  return result;
-}
-
-// TODO: this function
-// Handle aggregators
-std::pair<std::vector<AstArgument*>, bool> getInlinedArgument(AstProgram& program, const AstArgument* arg) {
-  bool changed = false;
-  std::vector<AstArgument*> versions;
-
-  if(const AstAggregator* aggr = dynamic_cast<const AstAggregator*>(arg)){
-    // first check the argument
-    // TODO: how to handle this correctly? what does `sum aggr1 : aggr2` even mean?
-    if(aggr->getTargetExpression() != nullptr) {
-      std::pair<std::vector<AstArgument*>, bool> argumentVersions = getInlinedArgument(program, aggr->getTargetExpression());
-      if(argumentVersions.second) {
-        changed = true;
-        // TODO: add in the cases for MAX, MIN, COUNT, SUM (all aggr types)
-        for(AstArgument* newArg : argumentVersions.first) {
-          AstAggregator* newAggr = new AstAggregator(aggr->getOperator());
-          newAggr->setTargetExpression(std::unique_ptr<AstArgument>(newArg));
-          for(AstLiteral* lit : aggr->getBodyLiterals()) {
-            newAggr->addBodyLiteral(std::unique_ptr<AstLiteral>(lit->clone()));
-          }
-          versions.push_back(newAggr);
-        }
-      }
-    }
-
-    // handle the body arguments
-    // TODO: only handle one change at a time for now?
-    if(!changed) {
-      std::vector<AstLiteral*> bodyLiterals = aggr->getBodyLiterals();
-      for(int i = 0; i < bodyLiterals.size(); i++) {
-        AstLiteral* currLit = bodyLiterals[i];
-        auto literalVersions = getInlinedLiteral(program, currLit);
-
-        // literal is inlined
-        if(literalVersions.first.second) {
-          changed = true;
-          AstAggregator::Op op = aggr->getOperator();
-
-          std::vector<AstArgument*> aggrVersions;
-
-          for(std::vector<AstLiteral*> inlineVersions : literalVersions.first.first) {
-            AstAggregator* newAggr = new AstAggregator(aggr->getOperator());
-            if(aggr->getTargetExpression() != nullptr) {
-              newAggr->setTargetExpression(std::unique_ptr<AstArgument>(aggr->getTargetExpression()->clone()));
-            }
-
-            // Add in everything except the current literal
-            for(int j = 0; j < bodyLiterals.size(); j++) {
-              if(i != j) {
-                newAggr->addBodyLiteral(std::unique_ptr<AstLiteral>(bodyLiterals[j]->clone()));
-              }
-            }
-
-            // Add in everything new
-            for(AstLiteral* addedLit : inlineVersions) {
-              newAggr->addBodyLiteral(std::unique_ptr<AstLiteral>(addedLit));
-            }
-
-            aggrVersions.push_back(newAggr);
-          }
-
-          // create the actual overall aggregator
-          if(op == AstAggregator::min) {
-            // min x : { a(x) }. <=> min ( min x : { a1(x) }, min x : { a2(x) }, ... )
-            versions.push_back(combineAggregators(aggrVersions, BinaryOp::MIN));
-          } else if (op == AstAggregator::max) {
-            // max x : { a(x) }. <=> max ( max x : { a1(x) }, max x : { a2(x) }, ... )
-            versions.push_back(combineAggregators(aggrVersions, BinaryOp::MAX));
-          } else if (op == AstAggregator::count) {
-            // count : { a(x) }. <=> sum ( count : { a1(x) }, count : { a2(x) }, ... )
-            versions.push_back(combineAggregators(aggrVersions, BinaryOp::ADD));
-          } else if (op == AstAggregator::sum) {
-            // sum x : { a(x) }. <=> sum ( sum x : { a1(x) }, sum x : { a2(x) }, ... )
-            versions.push_back(combineAggregators(aggrVersions, BinaryOp::ADD));
-          } else {
-            assert(false && "unhandled aggregator operator type");
-          }
-
-        } else if (literalVersions.second.second) {
-          changed = true;
-          for(AstLiteral* newLit : literalVersions.second.first) {
-            AstAggregator* newAggr = new AstAggregator(aggr->getOperator());
-            if(aggr->getTargetExpression() != nullptr) {
-              newAggr->setTargetExpression(std::unique_ptr<AstArgument>(aggr->getTargetExpression()->clone()));
-            }
-            for(int j = 0; j < bodyLiterals.size(); j++) {
-              if(i == j) {
-                newAggr->addBodyLiteral(std::unique_ptr<AstLiteral>(newLit));
-              } else {
-                newAggr->addBodyLiteral(std::unique_ptr<AstLiteral>(bodyLiterals[j]->clone()));
-              }
-            }
-            versions.push_back(newAggr);
-          }
-        }
-
-        // TODO: one at a time for now?
-        if(changed) {
-          break;
-        }
-      }
-    }
-  } else if (dynamic_cast<const AstFunctor*>(arg)) {
-    if(const AstUnaryFunctor* functor = dynamic_cast<const AstUnaryFunctor*>(arg)) {
-      std::pair<std::vector<AstArgument*>, bool> argumentVersions = getInlinedArgument(program, functor->getOperand());
-      if (argumentVersions.second) {
-        changed = true;
-        for(AstArgument* newArg : argumentVersions.first) {
-          AstArgument* newFunctor = new AstUnaryFunctor(functor->getFunction(), std::unique_ptr<AstArgument>(newArg));
-          versions.push_back(newFunctor);
-        }
-      }
-    } else if (const AstBinaryFunctor* functor = dynamic_cast<const AstBinaryFunctor*>(arg)) {
-      std::pair<std::vector<AstArgument*>, bool> lhsVersions = getInlinedArgument(program, functor->getLHS());
-      if(lhsVersions.second) {
-        changed = true;
-        for(AstArgument* newLhs : lhsVersions.first) {
-          AstArgument* newFunctor = new AstBinaryFunctor(functor->getFunction(), std::unique_ptr<AstArgument>(newLhs), std::unique_ptr<AstArgument>(functor->getRHS()));
-          versions.push_back(newFunctor);
-        }
-      } else {
-        std::pair<std::vector<AstArgument*>, bool> rhsVersions = getInlinedArgument(program, functor->getRHS());
-        if(rhsVersions.second) {
-          changed = true;
-          for(AstArgument* newRhs : rhsVersions.first) {
-            AstArgument* newFunctor = new AstBinaryFunctor(functor->getFunction(), std::unique_ptr<AstArgument>(functor->getLHS()), std::unique_ptr<AstArgument>(newRhs));
-            versions.push_back(newFunctor);
-          }
-        }
-      }
-    } else if (const AstTernaryFunctor* functor = dynamic_cast<const AstTernaryFunctor*>(arg)) {
-      std::pair<std::vector<AstArgument*>, bool> leftVersions = getInlinedArgument(program, functor->getArg(0));
-      if(leftVersions.second) {
-        changed = true;
-        for(AstArgument* newLeft : leftVersions.first) {
-          AstArgument* newFunctor = new AstTernaryFunctor(functor->getFunction(), std::unique_ptr<AstArgument>(newLeft), std::unique_ptr<AstArgument>(functor->getArg(1)->clone()), std::unique_ptr<AstArgument>(functor->getArg(2)->clone()));
-          versions.push_back(newFunctor);
-        }
-      } else {
-        std::pair<std::vector<AstArgument*>, bool> middleVersions = getInlinedArgument(program, functor->getArg(1));
-        if(middleVersions.second) {
-          changed = true;
-          for(AstArgument* newMiddle : middleVersions.first) {
-            AstArgument* newFunctor = new AstTernaryFunctor(functor->getFunction(), std::unique_ptr<AstArgument>(functor->getArg(0)->clone()), std::unique_ptr<AstArgument>(newMiddle), std::unique_ptr<AstArgument>(functor->getArg(2)->clone()));
-            versions.push_back(newFunctor);
-          }
-        } else {
-          std::pair<std::vector<AstArgument*>, bool> rightVersions = getInlinedArgument(program, functor->getArg(1));
-          if(rightVersions.second) {
-            changed = true;
-            for(AstArgument* newRight : rightVersions.first) {
-              AstArgument* newFunctor = new AstTernaryFunctor(functor->getFunction(), std::unique_ptr<AstArgument>(functor->getArg(0)->clone()), std::unique_ptr<AstArgument>(functor->getArg(1)->clone()), std::unique_ptr<AstArgument>(newRight));
-              versions.push_back(newFunctor);
-            }
-          }
-        }
-      }
-    }
-  } else if (const AstTypeCast* cast = dynamic_cast<const AstTypeCast*>(arg)) {
-    std::pair<std::vector<AstArgument*>, bool> argumentVersions = getInlinedArgument(program, cast->getValue());
-    if(argumentVersions.second) {
-      changed = true;
-      for(AstArgument* newArg : argumentVersions.first) {
-        AstArgument* newTypeCast = new AstTypeCast(std::unique_ptr<AstArgument>(newArg), cast->getType());
-        versions.push_back(newTypeCast);
-      }
-    }
-  } else if (const AstRecordInit* record = dynamic_cast<const AstRecordInit*>(arg)) {
-    std::vector<AstArgument*> recordArguments = record->getArguments();
-    for(int i = 0; i < recordArguments.size(); i++) {
-      AstArgument* currentRecArg = recordArguments[i];
-      std::pair<std::vector<AstArgument*>, bool> argumentVersions = getInlinedArgument(program, currentRecArg);
-      if(argumentVersions.second) {
-        changed = true;
-        for(AstArgument* newArgumentVersion : argumentVersions.first) {
-          AstRecordInit* newRecordArg = new AstRecordInit();
-          for(int j = 0; j < recordArguments.size(); j++) {
-            if (i == j) {
-              newRecordArg->add(std::unique_ptr<AstArgument>(newArgumentVersion));
-            } else {
-              newRecordArg->add(std::unique_ptr<AstArgument>(recordArguments[j]->clone()));
-            }
-          }
-          versions.push_back(newRecordArg);
-        }
-      }
-
-      // TODO: for now, only one change at a time
-      if(changed) {
-        break;
-      }
-    }
-  }
-
-  return std::make_pair(versions, changed);
-}
-
-std::pair<std::vector<AstAtom*>, bool> getInlinedAtom(AstProgram& program, AstAtom& atom) {
-  bool changed = false;
-  std::vector<AstAtom*> versions;
-
-  std::vector<AstArgument*> arguments = atom.getArguments();
-  for(int i = 0; i < arguments.size(); i++) {
-    AstArgument* arg = arguments[i];
-    std::pair<std::vector<AstArgument*>, bool> argumentVersions = getInlinedArgument(program, arg);
-    if(argumentVersions.second) {
-      changed = true;
-      for(AstArgument* newArgument : argumentVersions.first) {
-        AstAtom* newAtom = atom.clone();
-        newAtom->setArgument(i, std::unique_ptr<AstArgument>(newArgument));
-        versions.push_back(newAtom);
-      }
-    }
-
-    // TODO: for now, only one change at a time
-    if(changed) {
-      break;
-    }
-  }
-
-  return std::make_pair(versions, changed);
-}
-
-std::pair<std::vector<AstClause*>, bool> getInlinedClause(AstProgram& program, const AstClause& clause) {
-  bool changed = false;
-  std::vector<AstClause*> versions;
-
-  // Fix arguments in the head
-  AstAtom* head = clause.getHead();
-  auto headVersions = getInlinedAtom(program, *head);
-  if(headVersions.second) {
-    changed = true;
-    for(AstAtom* newHead : headVersions.first) {
-      AstClause* newClause = new AstClause();
-      newClause->setSrcLoc(clause.getSrcLoc());
-      newClause->setHead(std::unique_ptr<AstAtom>(newHead));
-      for(AstLiteral* lit : clause.getBodyLiterals()) {
-        newClause->addToBody(std::unique_ptr<AstLiteral>(lit->clone()));
-      }
-      versions.push_back(newClause);
-    }
-  }
-
-  // Go through the atoms in the clause and inline the necessary atoms
-  // TODO: this only does atoms... fix this up
-  if(!changed) {
-    auto bodyLiterals = clause.getBodyLiterals();
-    for(int i = 0; i < bodyLiterals.size(); i++) {
-      AstLiteral* currLit = bodyLiterals[i];
-      auto litVersions = getInlinedLiteral(program, currLit);
-      if(litVersions.first.second) {
-        changed = true;
-        for(auto body : litVersions.first.first) {
-          AstClause* replacementClause = clause.cloneHead();
-          for(AstLiteral* oldLit : bodyLiterals) {
-            if(currLit != oldLit) {
-              replacementClause->addToBody(std::unique_ptr<AstLiteral>(oldLit->clone()));
-            }
-          }
-
-          for(AstLiteral* newLit : body) {
-            replacementClause->addToBody(std::unique_ptr<AstLiteral>(newLit));
-          }
-
-          versions.push_back(replacementClause);
-        }
-      } else if (litVersions.second.second) {
-        changed = true;
-        for(AstLiteral* newLit : litVersions.second.first) {
-          AstClause* replacementClause = clause.cloneHead();
-          for(AstLiteral* oldLit : bodyLiterals) {
-            if(currLit != oldLit) {
-              replacementClause->addToBody(std::unique_ptr<AstLiteral>(oldLit->clone()));
-            }
-          }
-          replacementClause->addToBody(std::unique_ptr<AstLiteral>(newLit));
-
-          versions.push_back(replacementClause);
-        }
-      }
-
-      if(changed) {
-        // To avoid confusion, only replace one atom per iteration
-        // TODO: change this up later to make Smarter (tm)
-        break;
-      }
-    }
-  }
-
-  return std::make_pair(versions, changed);
-}
-
-void normaliseInlinedConstraints(AstProgram& program) {
-  static int newVarCount = 0;
-  for(AstRelation* rel : program.getRelations()) {
-    if (rel->isInline()) {
-      for(AstClause* clause : rel->getClauses()) {
-        AstClause* newClause = new AstClause();
-        newClause->setSrcLoc(clause->getSrcLoc());
-
-        AstAtom* clauseHead = new AstAtom();
-        clauseHead->setName(clause->getHead()->getName());
-        for(AstArgument* arg : clause->getHead()->getArguments()) {
-          if(AstConstant* constant = dynamic_cast<AstConstant*>(arg)) {
-            std::stringstream newVar;
-            newVar << "<new_var_" << newVarCount++ << ">" << std::endl;
-            clauseHead->addArgument(std::unique_ptr<AstArgument>(new AstVariable(newVar.str())));
-            newClause->addToBody(std::unique_ptr<AstLiteral>(new AstConstraint(
-                    BinaryConstraintOp::EQ, std::unique_ptr<AstArgument>(new AstVariable(newVar.str())),
-                    std::unique_ptr<AstArgument>(constant->clone()))));
-          } else {
-            clauseHead->addArgument(std::unique_ptr<AstArgument>(arg->clone()));
-          }
-        }
-
-        newClause->setHead(std::unique_ptr<AstAtom>(clauseHead));
-        for(AstLiteral* lit : clause->getBodyLiterals()) {
-          newClause->addToBody(std::unique_ptr<AstLiteral>(lit->clone()));
-        }
-        rel->addClause(std::unique_ptr<AstClause>(newClause));
-        program.removeClause(clause);
-      }
-    }
-  }
-}
-
-// TODO: are constraints handled properly?
-bool InlineRelationsTransformer::transform(AstTranslationUnit& translationUnit) {
-  bool changed = false;
-
-  AstProgram& program = *translationUnit.getProgram();
-
-  normaliseInlinedConstraints(program);
-
-  renameInlinedArguments(program);
-
-  nameUnderscores(program);
-
-  // Keep trying to inline until we reach a fixed point
-  // TODO: why is this important?
-  bool clausesChanged = true;
-  while(clausesChanged) {
-    std::vector<const AstClause*> clausesToDelete;
-    clausesChanged = false;
-
-    // Go through each clause in the program and check if we need to inline anything
-    visitDepthFirst(program, [&](const AstClause& clause) {
-      // Skip if the clause is meant to be inlined
-      // TODO: should this really be done? is it more efficient if kept?
-      if(program.getRelation(clause.getHead()->getName())->isInline()) {
-        return;
-      }
-
-      // Get the inlined versions of this clause
-      std::pair<std::vector<AstClause*>, bool> newClauses = getInlinedClause(program, clause);
-
-      if(newClauses.second) {
-        // Clause was replaced with equivalent versions, so delete it!
-        clausesToDelete.push_back(&clause);
-
-        // Add in the new clauses
-        for(AstClause* replacementClause : newClauses.first) {
-          program.appendClause(std::unique_ptr<AstClause>(replacementClause));
-        }
-
-        clausesChanged = true;
-        changed = true;
-      }
-    });
-
-    // Delete all clauses that were replaced
-    // TODO: double check positioning of clause deletion
-    for(const AstClause* clause : clausesToDelete) {
-      program.removeClause(clause);
-    }
-  }
-
-  //unnameUnderscores(program);
-
-  std::cout << program << std::endl;
-  return changed;
 }
 
 bool NormaliseConstraintsTransformer::transform(AstTranslationUnit& translationUnit) {
